# metal : use residency sets