fix #10119 

~Investigating further if the issue described in #10119 can be fixed by using [MTLResidencySet](https://developer.apple.com/documentation/metal/mtlresidencyset?language=objc). Unfortunately, using the residency sets does not make a difference. I tried both attaching them to the command queue and to the command buffers, but the overhead after 1 second of being idle remains. Not sure if I am using it in a wrong way, or this is not related to the wired memory collector logic controlled by `sysctl iogpu.disable_wired_collector`.~

Ignore the above - I was measuring incorrectly.

Using residency sets makes the allocated memory stay wired and eliminates almost completely the overhead observed in #10119. For example, on M2 Ultra, using 7B Q8_0 model the requests are ~250ms faster thanks to this change. It seems it is not necessary to attach the residency sets to the command queue and buffers, so the change is rather simple. For each buffer, we create an associated `MTLResidencySet` and add the `MTLBuffer` objects to it. After that we commit it and request residency:

https://github.com/ggerganov/llama.cpp/blob/225d2e0ca1d7a7e627f2cea4a43dd77a83b9f078/ggml/src/ggml-metal/ggml-metal.m#L1084-L1091

build: b9126fe36 (4561)

| Model         | Test   |   t/s master |   t/s gg/metal-residency-sets |   Speedup |
|:--------------|:-------|-------------:|------------------------------:|----------:|
| llama 3B F16  | pp512  |      3289.51 |                       3286.29 |      1.00 |
| llama 3B F16  | tg128  |        73.28 |                         73.35 |      1.00 |
| llama 3B Q4_0 | pp512  |      2999.71 |                       3002.93 |      1.00 |
| llama 3B Q4_0 | tg128  |       165.83 |                        166.03 |      1.00 |
| llama 3B Q8_0 | pp512  |      2958.32 |                       2960.69 |      1.00 |
| llama 3B Q8_0 | tg128  |       123.61 |                        123.96 |      1.00 |


## Metal backend changes

Checks the environment variable `GGML_METAL_NO_RESIDENCY`. If set, then no residency sets will be created, allowing the GPU memory to be collected by the OS after 1 second of inactivity. Generally, this should rarely be needed as it hurts the performance of the application, but keeping support just in case.