@ggerganov 2025-01-26
Great news - this change finally resolves the annoying overhead that I was observing. The only remaining question is how to implement this to be compatible with macOS < 15.0.

Any suggestions?